Fix for setting rtc compare when timeout just occurred.
In us_ticker.c, need to make sure that we are not messing up the compare
register calculation when the timeout just occurred.  In previous code,
could add 2^32 before calculating the rtc compare register when the
timeout occurred.  That results in the compare register set to 199
seconds late.  In the fix, just force rtc to expire as soon as
possible.

The problem can occur when mbed_ticker_api.c calls set_interrupt in
ticker_irq_handler().  It calls it when the timeout is >0.  However, a tick
could occur in the rtc before the code in us_ticker.c gets the request.
If the timeout is really small, the expiration time could be before the
current time.